Move plot_stacked_histogram to separate file.
