+++ ms2deepscore/wrapper_functions/plotting_wrapper_functions.py
@@ -10,8 +10,7 @@ from ms2deepscore.train_new_model.SettingMS2Deepscore import \
 from ms2deepscore.utils import load_pickled_file, save_pickled_file
 from ms2deepscore.visualize_results.calculate_tanimoto_scores import \
     get_tanimoto_score_between_spectra
-from ms2deepscore.visualize_results.plotting import \
-    plot_stacked_histogram_plot_wrapper
+from ms2deepscore.visualize_results.plot_stacked_histogram import plot_stacked_histogram_plot_wrapper
 
 
 class BenchmarkingResultsFileNames:
